backend/test/integration/requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
attrs==22.1.0
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.0.4
idna==3.4
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.2
platformdirs==2.6.0
pluggy==1.0.0
PyHamcrest==2.0.4
pyparsing==3.0.9
pytest==7.2.0
requests==2.28.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
